+++ packages/components/CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 #### Changed
 
 - Improve height calculation for `Accordion`
+- Improve slotted anchor support for `Link Pure`
 
 ### [2.17.0] - 2022-10-31